+++ src/main/java/kopo/poly/controller/IndexController.java
@@ -14,18 +14,6 @@ public class IndexController {
         log.info(this.getClass().getName()+".index End!!");
         return "index";
     }
-    @GetMapping(value = "login")
-    public String login() throws Exception {
-        log.info(this.getClass().getName()+".login Start!!");
-        log.info(this.getClass().getName()+".login End!!");
-        return "login";
-    }
 
-    @GetMapping(value = "register")
-    public String register() throws Exception {
-        log.info(this.getClass().getName()+".register Start!!");
-        log.info(this.getClass().getName()+".register End!!");
-        return "register";
-    }
 
 }
