+++ src/main/webapp/WEB-INF/views/login/find-pw.jsp
@@ -1,0 +1,11 @@
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>웹페이지 제목</title>
+</head>
+<body>
+
+</body>
+</html>
